# swagger.yaml
openapi: 3.0.3
info:
  title: Алимбаев Талгат red_mad_robot
  description: |-
    Привет, тут я собрал описание API, которые используются в моем решении тестового задания
  contact:
    email: alimbayev.talat@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/bids/create:
    post:
      tags:
        - bids
      summary: создаем новую ставку
      description: создание новой ставки доступно только для зарегистрированных пользователей
      operationId: createBid
      requestBody:
        description: создание новой ставки
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Bid'
        required: true
      responses:
        '201':
          description: ставка создана
        '400':
          description: При вводе некорректной информации выходит статус 400
...
  /api/bids/by-ad:
    get:
      tags:
        - bids
      summary: поиск всех ставок по объявлению
      description: поиск всех ставок по объявлению
      operationId: bidsByAd
      parameters: 
        - name: adId
          in: query
          description: номер объявления, по которому осуществляется поиск всех ставок
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: список всех ставок по данному объявлению
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bids'
        '404':
          description: такое объявление не существует
  /api/bids/highest-bid/by-ad:
    get:
      tags:
        - bids
      summary: поиск наивысшей ставки по объявлению
      description: поиск наивысшей ставки по объявлению
      operationId: highestBidByAd
      parameters: 
        - name: adId
          in: query
          description: номер объявления, по которому осуществляется поиск наивысшей ставки
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: наивысшая ставка по данному объявлению
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidFull'
        '404':
          description: такое объявление не существует
...
components:
  schemas:
    Bid:
      type: object
      properties:
        ad:
          type: object
          properties:
            id:
              type: integer
              example: 1
        bid:
          type: number
          example: 150.5
        user:
          type: object
          properties:
            id:
              type: integer
              example: 1
    BidFull:
      type: object
      properties:
        id:
          type: integer
          example: 1
        ad:
          $ref: '#/components/schemas/AdFull'
        bid:
          type: number
          example: 1234
        user:
          $ref: '#/components/schemas/User'
    Bids:
      type: array
      items:
        oneOf:
          - $ref: '#/components/schemas/BidFull'
          - $ref: '#/components/schemas/BidFull'
...
    AdFull:
      type: object
      properties:
        id:
          type: integer
          example: 1
        description:
          type: string
          example: мое первое объявление
        startPrice:
          type: number
          example: 150.5
        imageLink:
          type: string
          example: https://images.google.com/
        adDurationMinutes:
          type: integer
          example: 1
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
...